+++ modules/web/web-app/lib/schemas/index.ts
@@ -2,6 +2,7 @@ export * as addons from './addons'
 export * as billableMetrics from './billableMetrics'
 export * as coupons from './coupons'
 export * as customers from './customers'
+export * as invitation from './invitation'
 export * as me from './me'
 export * as organizations from './organizations'
 export * as products from './products'
