+++ modules/web/web-app/lib/schemas/invitation.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod'
+
+export const invitationSchema = z.object({
+  email: z.array(z.string().email()),
+})
