+++ modules/web/web-app/pages/onboarding/index.ts
@@ -1,2 +1,3 @@
+export { InviteOnboarding } from './inviteOnboarding'
 export { OrganizationOnboarding } from './organizationOnboarding'
 export { UserOnboarding } from './userOnboarding'
